requirements.txt
pytest-watch==4.2.0
pytest==7.1.3
pillow==9.2.0
google-cloud-vision
google-cloud-translate==2.0.0